@@ -18,84 +18,95 @@
   "main": "main.js",
   "private": true,
   "scripts": {
-    "postinstall": "yarn postinstall:electron && electron-builder install-app-deps",
-    "postinstall:web": "node postinstall-web",
-    "postinstall:electron": "node postinstall",
+    "postinstall": "electron-builder install-app-deps",
     "ng": "ng",
-    "start": "yarn postinstall:electron && npm-run-all -p ng:serve electron:serve",
-    "build": "yarn postinstall:electron && yarn electron:serve-tsc && ng build",
-    "build:dev": "yarn build -c dev",
-    "build:prod": "yarn build -c production",
+    "start": "npm-run-all -p electron:serve ng:serve",
+    "build": "npm run electron:serve-tsc && ng build --base-href ./",
+    "build:dev": "npm run build -- -c dev",
+    "build:prod": "npm run build -- -c production",
     "ng:serve": "ng serve",
-    "ng:serve:web": "yarn postinstall:web && ng serve -o",
     "electron:serve-tsc": "tsc -p tsconfig-serve.json",
-    "electron:serve": "wait-on http-get://localhost:4200/ && yarn electron:serve-tsc && electron . --serve",
-    "release:linux": "yarn build:prod && electron-builder build --linux",
-    "release:windows": "yarn build:prod && electron-builder build --windows",
-    "release:mac": "yarn build:prod && electron-builder build --mac",
-    "release:all": "yarn build:prod && electron-builder build --windows --linux --mac",
-    "publish:linux": "yarn build:prod && electron-builder build --linux -p always",
-    "publish:windows": "yarn build:prod && electron-builder build --windows -p always",
-    "publish:all": "yarn build:prod && electron-builder build --windows --linux --mac -p always"
+    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:serve-tsc && electron . --serve",
+    "electron:local": "npm run build:prod && electron .",
+    "electron:linux": "npm run build:prod && electron-builder build --linux",
+    "electron:windows": "npm run build:prod && electron-builder build --windows",
+    "electron:mac": "npm run build:prod && electron-builder build --mac",
+    "test": "ng test",
+    "e2e": "npm run build:prod && cross-env TS_NODE_PROJECT='e2e/tsconfig.e2e.json' mocha --timeout 300000 --require ts-node/register e2e/**/*.e2e.ts",
+    "version": "conventional-changelog -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md",
+    "lint": "ng lint",
+    "release:linux": "npm run build:prod && npm run electron-builder build --linux",
+    "release:windows": "npm run build:prod && npm run electron-builder build --windows",
+    "release:mac": "npm run build:prod && electron-builder build --mac",
+    "release:all": "npm run build:prod && electron-builder build --windows --linux --mac",
+    "publish:linux": "npm run build:prod && electron-builder build --linux -p always",
+    "publish:windows": "npm run build:prod && electron-builder build --windows -p always",
+    "publish:mac": "npm run build:prod && electron-builder build --mac -p always",
+    "publish:all": "npm run build:prod && electron-builder build --windows --linux --mac -p always"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.1",
-    "@angular/animations": "^9.0.0",
-    "@angular/cli": "9.0.1",
-    "@angular/common": "9.0.0",
-    "@angular/compiler": "9.0.0",
-    "@angular/compiler-cli": "9.0.0",
-    "@angular/core": "9.0.0",
-    "@angular/forms": "9.0.0",
-    "@angular/language-service": "9.0.0",
-    "@angular/localize": "^9.0.0",
-    "@angular/platform-browser": "9.0.0",
-    "@angular/platform-browser-dynamic": "9.0.0",
-    "@angular/router": "9.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^5.1.1",
-    "@ngx-translate/core": "12.0.0",
+    "@angular-builders/custom-webpack": "^9.1.0",
+    "@angular-devkit/build-angular": "^0.900.7",
+    "@angular-eslint/builder": "^0.0.1-alpha.29",
+    "@angular/animations": "^9.1.4",
+    "@angular/cli": "^9.1.4",
+    "@angular/common": "9.1.4",
+    "@angular/compiler": "9.1.4",
+    "@angular/compiler-cli": "9.1.4",
+    "@angular/core": "9.1.4",
+    "@angular/forms": "9.1.4",
+    "@angular/language-service": "9.1.4",
+    "@angular/localize": "^9.1.4",
+    "@angular/platform-browser": "9.1.4",
+    "@angular/platform-browser-dynamic": "9.1.4",
+    "@angular/router": "9.1.4",
+    "@ng-bootstrap/ng-bootstrap": "^6.1.0",
+    "@ngx-translate/core": "12.1.2",
     "@ngx-translate/http-loader": "4.0.0",
-    "@types/lodash": "^4.14.139",
-    "@types/node": "12.12.6",
+    "@types/lodash": "^4.14.150",
+    "@types/node": "12.12.21",
     "@types/request-promise-native": "^1.0.17",
-    "apollo-angular": "^1.7.0",
-    "apollo-angular-link-http": "^1.8.0",
-    "apollo-cache-inmemory": "^1.6.3",
-    "apollo-client": "^2.6.4",
-    "apollo-link": "^1.2.13",
-    "apollo-link-retry": "^2.2.15",
-    "bootstrap-scss": "^4.3.1",
-    "clappr": "^0.3.9",
-    "core-js": "3.6.4",
-    "electron": "8.0.0",
-    "electron-builder": "22.3.2",
+    "@typescript-eslint/eslint-plugin": "^2.30.0",
+    "@typescript-eslint/parser": "^2.30.0",
+    "apollo-angular": "^1.9.1",
+    "apollo-angular-link-http": "^1.10.0",
+    "apollo-cache-inmemory": "^1.6.5",
+    "apollo-client": "^2.6.8",
+    "apollo-link": "^1.2.14",
+    "apollo-link-retry": "^2.2.16",
+    "bootstrap-scss": "^4.4.1",
+    "clappr": "^0.3.12",
+    "core-js": "3.6.5",
+    "electron": "8.2.5",
+    "electron-builder": "22.6.0",
     "electron-reload": "1.5.0",
-    "electron-store": "^5.0.0",
-    "graphql": "^14.5.8",
-    "graphql-tag": "^2.10.1",
+    "electron-store": "^5.1.1",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "graphql": "^14.0.0",
+    "graphql-tag": "^2.10.3",
     "level-selector": "^0.2.0",
     "lodash": "^4.17.15",
     "material-icons": "^0.3.1",
-    "moment": "^2.24.0",
     "normalize-scss": "^7.0.1",
-    "npm-run-all": "4.1.5",
-    "prettier": "^1.19.1",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.0.5",
+    "prettier-eslint": "^9.0.1",
     "roboto-fontface": "^0.10.0",
-    "rxjs": "6.5.4",
+    "rxjs": "6.5.5",
     "simplebar": "^5.1.0",
-    "ts-node": "8.6.2",
-    "tslint": "6.0.0",
-    "tslint-config-prettier": "^1.18.0",
-    "typescript": "3.7.5",
-    "wait-on": "4.0.0",
+    "ts-node": "8.9.1",
+    "typescript": "3.8.3",
+    "wait-on": "4.0.2",
     "webdriver-manager": "13.0.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.10.3"
   },
   "dependencies": {
-    "electron-log": "^4.0.6",
-    "electron-updater": "^4.1.2",
-    "request": "^2.88.0",
-    "request-promise-native": "^1.0.7",
-    "tslib": "^1.10.0"
+    "electron-log": "^4.1.2",
+    "electron-updater": "^4.3.1",
+    "request": "^2.88.2",
+    "request-promise-native": "^1.0.8",
+    "tslib": "^1.11.1"
   }
 }
