@@ -1,10 +1,10 @@
-import {Component, NgZone, OnInit} from '@angular/core';
-import {ActivatedRoute, Router} from '@angular/router';
-import {ChannelService, IStream} from '../../providers/channels.service';
-import {ErrorService} from '../../providers/errorhandler.service';
-import {GameService, IGame} from '../../providers/games.service';
-import {SpinnerService} from '../../providers/spinner.service';
-import {ToolbarService} from '../../providers/toolbar.service';
+import { Component, NgZone, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+import { ChannelService, Stream } from '../../providers/channels.service';
+import { ErrorService } from '../../providers/errorhandler.service';
+import { GameService, Game } from '../../providers/games.service';
+import { SpinnerService } from '../../providers/spinner.service';
+import { ToolbarService } from '../../providers/toolbar.service';
 
 @Component({
   templateUrl: './channels.component.html',
@@ -12,8 +12,8 @@ import {ToolbarService} from '../../providers/toolbar.service';
   styleUrls: ['./channels.component.scss']
 })
 export class ChannelsComponent implements OnInit {
-  private game: IGame = null;
-  streams: IStream[] = [];
+  private game: Game = null;
+  streams: Stream[] = [];
   fetchingMore = false;
 
   constructor(
@@ -27,8 +27,8 @@ export class ChannelsComponent implements OnInit {
     private zone: NgZone
   ) {}
 
-  ngOnInit() {
-    this.route.params.subscribe(params => {
+  ngOnInit(): void {
+    this.route.params.subscribe((params) => {
       this.spinnerService.show();
 
       if (params['game'] !== 'top' && params['game'] !== 'following') {
@@ -38,7 +38,7 @@ export class ChannelsComponent implements OnInit {
       if (params['game'] === 'top') {
         this.toolbarService.setTitle('Top Streams');
         this.fetchingMore = true;
-        this.channelService.getTopStreams().then((streams: IStream[]) => {
+        this.channelService.getTopStreams().then((streams: Stream[]) => {
           this.streams = streams;
           this.fetchingMore = false;
           this.spinnerService.hide();
@@ -46,7 +46,7 @@ export class ChannelsComponent implements OnInit {
       } else if (params['game'] === 'following') {
         this.toolbarService.setTitle('Followed Streams');
         this.fetchingMore = true;
-        this.channelService.getFollowedStreams().then((streams: IStream[]) => {
+        this.channelService.getFollowedStreams().then((streams: Stream[]) => {
           this.streams = streams;
           this.fetchingMore = false;
           this.spinnerService.hide();
@@ -54,11 +54,13 @@ export class ChannelsComponent implements OnInit {
       } else if (this.game) {
         this.toolbarService.setTitle(this.game.name);
         this.fetchingMore = true;
-        this.channelService.getGameStreams(this.game).then((streams: IStream[]) => {
-          this.streams = streams;
-          this.fetchingMore = false;
-          this.spinnerService.hide();
-        });
+        this.channelService
+          .getGameStreams(this.game)
+          .then((streams: Stream[]) => {
+            this.streams = streams;
+            this.fetchingMore = false;
+            this.spinnerService.hide();
+          });
       }
 
       // Set toolbar icon
@@ -66,25 +68,25 @@ export class ChannelsComponent implements OnInit {
     });
   }
 
-  itemClicked(stream: IStream) {
+  itemClicked(stream: Stream): void {
     this.channelService.currentStream = stream;
     this.router.navigate(['/play/' + stream.id]);
   }
 
   // Triggered when stream list is scrolled to the bottom (infinite-scroll)
-  onScrolled() {
+  onScrolled(): void {
     // Load more items only if we are not already doing that
     if (!this.fetchingMore) {
       this.fetchingMore = true;
       this.zone.run(() => {});
       this.channelService
         .fetchMoreStreams()
-        .then((streams: IStream[]) => {
+        .then((streams: Stream[]) => {
           this.streams = streams;
           this.fetchingMore = false;
           this.zone.run(() => {});
         })
-        .catch(reason => {
+        .catch((reason) => {
           console.log(reason);
           this.fetchingMore = false;
           this.zone.run(() => {});
