@@ -1,12 +1,11 @@
 import { Component, NgZone, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Router } from '@angular/router';
 import { concat as _concat } from 'lodash';
 
 import { ErrorService } from '../../providers/errorhandler.service';
 import { GameService } from '../../providers/games.service';
 import { SpinnerService } from '../../providers/spinner.service';
 import { ToolbarService } from '../../providers/toolbar.service';
-import { TwitchService } from '../../providers/twitch.service';
 
 @Component({
   templateUrl: './games.component.html',
@@ -25,7 +24,7 @@ export class GamesComponent implements OnInit {
     private zone: NgZone
   ) {}
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.spinnerService.show();
 
     // Set toolbar tile and logo
@@ -39,7 +38,7 @@ export class GamesComponent implements OnInit {
         this.games = _concat(this.games, games);
         this.spinnerService.hide();
       })
-      .catch(reason => {
+      .catch((reason) => {
         this.spinnerService.hide();
         this.errorService.showError('Failed fetching games', reason);
         console.log(reason);
@@ -47,7 +46,7 @@ export class GamesComponent implements OnInit {
   }
 
   // Triggered when list is scrolled to bottom (ininite-scroll)
-  onScrolled() {
+  onScrolled(): void {
     // Only fetch more if we are not already doing that
     if (!this.fetchingMore) {
       this.fetchingMore = true;
@@ -60,7 +59,7 @@ export class GamesComponent implements OnInit {
           this.fetchingMore = false;
           this.zone.run(() => {});
         })
-        .catch(reason => {
+        .catch((reason) => {
           console.log(reason);
           this.fetchingMore = false;
           this.zone.run(() => {});
