@@ -3,7 +3,7 @@ import { ActivatedRoute } from '@angular/router';
 
 import Clappr from 'clappr';
 import LevelSelector from 'level-selector';
-import { ChannelService, IStream } from '../../providers/channels.service';
+import { ChannelService, Stream } from '../../providers/channels.service';
 import { ErrorService } from '../../providers/errorhandler.service';
 import { SettingsService } from '../../providers/settings.service';
 import { ToolbarService } from '../../providers/toolbar.service';
@@ -16,8 +16,8 @@ import { TwitchService } from '../../providers/twitch.service';
   styleUrls: ['./player.component.scss']
 })
 export class PlayerComponent implements OnInit, OnDestroy {
-  stream: IStream;
-  chat_url: string;
+  stream: Stream;
+  chatUrl: string;
   player: any;
   isLoading = true;
   player_source_resolution = 0;
@@ -31,13 +31,13 @@ export class PlayerComponent implements OnInit, OnDestroy {
     private errorService: ErrorService
   ) {}
 
-  async ngOnInit() {
-    this.route.params.subscribe(params => {
+  async ngOnInit(): Promise<void> {
+    this.route.params.subscribe(() => {
       this.stream = this.channelService.currentStream;
     });
 
     if (this.stream.broadcaster && this.stream.broadcaster.login) {
-      this.chat_url = `https://www.twitch.tv/embed/${this.stream.broadcaster.login}/chat?darkpopout`;
+      this.chatUrl = `https://www.twitch.tv/embed/${this.stream.broadcaster.login}/chat?darkpopout`;
     }
 
     // Set toolbar title and logo
@@ -48,13 +48,13 @@ export class PlayerComponent implements OnInit, OnDestroy {
 
     // Set toolbar subheader info
     this.toolbarService.setSubheader({
-      player_username:
+      playerUsername:
         this.stream.broadcaster && this.stream.broadcaster.displayName,
-      player_game:
+      playerGame:
         this.stream.broadcaster &&
         this.stream.broadcaster.broadcastSettings &&
         this.stream.broadcaster.broadcastSettings.game.name,
-      player_logo: '' // FIXME
+      playerLogo: '' // FIXME
     });
 
     let sourceUrl;
@@ -72,7 +72,7 @@ export class PlayerComponent implements OnInit, OnDestroy {
       },
       levelSelectorConfig: {
         title: 'Quality',
-        labelCallback: (playbackLevel, customLabel) => {
+        labelCallback: (playbackLevel): string => {
           return (
             playbackLevel.level.height +
             'p' +
@@ -87,7 +87,7 @@ export class PlayerComponent implements OnInit, OnDestroy {
 
       playback: {
         hlsjsConfig: {
-          maxMaxBufferLength: this.settingsService.getConfig().buffer_length,
+          maxMaxBufferLength: this.settingsService.getConfig().bufferLength,
           liveSyncDurationCount: 2,
           liveMaxLatencyDurationCount: 3
         }
@@ -97,28 +97,28 @@ export class PlayerComponent implements OnInit, OnDestroy {
     this.player.listenToOnce(
       this.player.core.activePlayback,
       Clappr.Events.PLAYBACK_LEVELS_AVAILABLE,
-      levels => {
-        if (this.settingsService.getConfig().preferred_quality !== 'auto') {
-          let target_quality_id = -1;
+      (levels) => {
+        if (this.settingsService.getConfig().preferredQuality !== 'auto') {
+          let targetQualityId = -1;
 
-          if (this.settingsService.getConfig().preferred_quality === 'source') {
-            target_quality_id = levels[levels.length - 1].id;
+          if (this.settingsService.getConfig().preferredQuality === 'source') {
+            targetQualityId = levels[levels.length - 1].id;
           } else {
-            levels.forEach(level => {
+            levels.forEach((level) => {
               if (
                 level.label.includes(
-                  this.settingsService.getConfig().preferred_quality
+                  this.settingsService.getConfig().preferredQuality
                 )
               ) {
-                target_quality_id = level.id;
+                targetQualityId = level.id;
               }
             });
           }
 
-          this.player.getPlugin('hls').currentLevel = target_quality_id;
+          this.player.getPlugin('hls').currentLevel = targetQualityId;
           this.player.getPlugin(
             'level_selector'
-          ).selectedLevelId = target_quality_id;
+          ).selectedLevelId = targetQualityId;
         }
       }
     );
@@ -128,12 +128,12 @@ export class PlayerComponent implements OnInit, OnDestroy {
     });
   }
 
-  ngOnDestroy() {
+  ngOnDestroy(): void {
     // Hide the toolbar subheader on component destroy
     this.toolbarService.setSubheader({
-      player_username: null,
-      player_game: null,
-      player_logo: null
+      playerUsername: null,
+      playerGame: null,
+      playerLogo: null
     });
 
     // Clear the player
