@@ -1,14 +1,9 @@
 import { Injectable } from '@angular/core';
 import { ApolloQueryResult } from 'apollo-client';
-import {
-  concat as _concat,
-  find as _find,
-  map as _map,
-  uniqBy as _uniqBy
-} from 'lodash';
-import { GetTopGamesGQL, ITopGamesResponse } from './twitch-graphql.service';
+import { uniqBy } from 'lodash';
+import { GetTopGamesGQL, TopGamesResponse } from './twitch-graphql.service';
 
-export interface IGame {
+export interface Game {
   id: string;
   name: string;
   boxArtURL: string;
@@ -18,19 +13,19 @@ export interface IGame {
 // Service that allows components to get game list information
 @Injectable()
 export class GameService {
-  private games: IGame[] = [];
+  private games: Game[] = [];
   private cursor = '';
 
   constructor(private getTopGamesGQL: GetTopGamesGQL) {}
 
-  async getTopGames() {
+  async getTopGames(): Promise<Game[]> {
     return new Promise((resolve, reject) => {
       this.getTopGamesGQL
         .fetch()
-        .subscribe((result: ApolloQueryResult<ITopGamesResponse>) => {
+        .subscribe((result: ApolloQueryResult<TopGamesResponse>) => {
           if (result.data) {
             this.cursor = result.data.games.edges.slice(-1)[0].cursor;
-            this.games = _map(result.data.games.edges, e => e.node);
+            this.games = result.data.games.edges.map((e) => e.node);
             resolve(this.games);
           } else {
             reject();
@@ -39,18 +34,15 @@ export class GameService {
     });
   }
 
-  async fetchMoreTopGames() {
+  async fetchMoreTopGames(): Promise<Game[]> {
     return new Promise((resolve, reject) => {
       this.getTopGamesGQL
         .fetch({ cursor: this.cursor })
-        .subscribe((result: ApolloQueryResult<ITopGamesResponse>) => {
+        .subscribe((result: ApolloQueryResult<TopGamesResponse>) => {
           if (result.data) {
             this.cursor = result.data.games.edges.slice(-1)[0].cursor;
-            this.games = _uniqBy(
-              _concat(
-                this.games,
-                _map(result.data.games.edges, e => e.node)
-              ),
+            this.games = uniqBy(
+              this.games.concat(result.data.games.edges.map((e) => e.node)),
               'id'
             );
 
@@ -62,7 +54,7 @@ export class GameService {
     });
   }
 
-  getGame(id: string) {
-    return _find(this.games, game => game.id === id);
+  getGame(id: string): Game {
+    return this.games.find((game) => game.id === id);
   }
 }
