@@ -1,4 +1,4 @@
 export const AppConfig = {
   production: true,
-  environment: "PROD"
+  environment: 'PROD'
 };
