@@ -12,4 +12,4 @@ platformBrowserDynamic()
   .bootstrapModule(AppModule, {
     preserveWhitespaces: false
   })
-  .catch(err => console.error(err));
+  .catch((err) => console.error(err));
